added initial configuration
