+++ demo/styles/default.st.css
@@ -1,7 +1,7 @@
 @namespace "demos";
 :import {
   -st-theme: true;
-  -st-from: "../src/themes/default/theme.st.css";
+  -st-from: "../../src/themes/default/theme.st.css";
   -st-named:  toggleSmall, btnLarge, btnSmall, btnTiny,
               btnSecondary,
               btnWhite, btnWhiteSecondary,
