+++ test/utils/with-theme.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {stylable} from 'wix-react-tools';
-import styles from '../../demo/style-default.st.css';
+import styles from '../../demo/styles/default.st.css';
 
 export const WithThemeDAID = 'THEMED_CONTAINER';
 export const WithTheme = (Node?: React.ReactNode, daid?: string, theme = styles): React.SFC => {
