+++ webpack.config.js
@@ -4,8 +4,7 @@ const stylableOptions = { injectBundleCss: true, nsDelimiter:'--' };
 module.exports = {
     devtool: 'source-map',
     entry: {
-        demos: ['core-js/shim', './demo/index-default.tsx'],
-        'demos-wix': ['core-js/shim', './demo/index-wix.tsx'],
+        demos: ['core-js/shim', './demo/index.tsx'],
         tests: ['core-js/shim', './test/utils/mobx.config.ts', 'mocha-loader!./test/webpack.ts']
     },
     module: {
